requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.17.0
fastapi==0.103.2
slack_bolt
openai==0.28.0
azure-storage-blob==12.18.3
openapi-schema-pydantic==1.2.4
azure-cosmos==4.5.1
langchain==0.0.340
llama-index==0.7.5
azure-search-documents==11.4.0b8
azure-common==1.1.28
azure-identity==1.13.0
python-dotenv==1.0.0
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
unstructured==0.5.6
pypdf==4.0.2
docx2txt==0.8
